requirements.txt
anki==2.1.35
ankirspy==2.1.35
aqt==2.1.35
attrs==20.3.0
beautifulsoup4==4.9.3
cached-property==1.5.2
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
decorator==4.4.2
Flask==1.1.2
Flask-Cors==3.0.9
frozendict==1.2
genanki==0.9.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
Markdown==3.3.3
MarkupSafe==1.1.1
mypy==0.790
mypy-extensions==0.4.3
orjson==3.4.3
protobuf==3.14.0
pyaml==20.4.0
PyQt5==5.15.1
PyQt5-sip==12.8.1
pyrsistent==0.17.3
PySocks==1.7.1
pystache==0.5.4
PyYAML==5.3.1
requests==2.25.0
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.0.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
waitress==1.4.4
Werkzeug==1.0.1
